Fixed incorrect pass when Kwalitee detected serious issues
According to https://metacpan.org/pod/App::CPANTS::Lint#lint

"Takes a path to a distribution tarball and analyses it. Returns true if the distribution has no significant issues (experimental metrics are always ignored). Otherwise, returns false."

Currently, if a check is done for Kwalitee, if no serious issues are found, it will die with message on line 28. This is the inverse of what is intended.